vector-db-2/requirements.txt
fastapi==0.115.2
pydantic==2.9.2
pymilvus==2.4.8
PyPDF2==3.0.1
PyYAML==6.0.2
PyYAML==6.0.2
sentence_transformers==3.2.0
uvicorn==0.32.0
